Merge EmbeddingsRequest into vanilla object before posting, rather than nesting it.
The third argument to `utils.post` should be the raw request body, but the `embeddings` method currently misses the spread operator used by all the other ollama wrappers. As a result, embeddings calls generate a JSON request body that looks like this:

```
{
  request: {
    model: 'llama2',
    prompt: 'Some text here...'
  }
}
```

instead of this:
```
{
  model: 'llama2',
  prompt: 'Some text here...'
}
```

This commit adds the spread operator, fixing the embeddings method.