chore: small improvement to tool example
The tool example was quite useful, but my first extension was to
add a second function which leads to this suggested change.

I think it is better to handle the mapping of the args object returned
to the specific argments needed in the function itself. This allows
additional functions to be added by simply adding to the json
desribing the available tools and the list of tools that can be called.

The example was already half way there by pulling the function from
an array based on the name, but missed handling the parameters in
a generic way to complete making invocation of the function generic.

Signed-off-by: Michael Dawson <midawson@redhat.com>